Move utils and tests
