+++ src/client/api/quoteCartQuerySelectors.ts
@@ -1,4 +1,3 @@
-import { getBundleVariantFromInsuranceTypesWithFallback } from 'pages/OfferNew/utils'
 import {
   QuoteCartQuery,
   QuoteBundleVariant,
@@ -7,6 +6,7 @@ import {
 } from 'data/graphql'
 import { InsuranceType } from 'utils/hooks/useSelectedInsuranceTypes'
 import { OfferData } from '../pages/Offer/types'
+import { getBundleVariantFromInsuranceTypesWithFallback } from '../pages/Offer/utils'
 
 export function getSelectedBundleVariant(
   quoteCartQuery: QuoteCartQuery | undefined,
